Improve identity request payload
Removes application specific request_key, adds ability to optionally specify requested permission level and shrinks the payload size to 1 byte for the most common request which is to ask for an unspecified id (i.e. id of current signer)

Also introduces a new placeholder. The name `............2` will now resolve to signing permission when used in actions