Add Authority type and refactor inter-dependencies
Unfortunately adding the Authority type was the straw that broke the camels back for the commonjs module resolution when handling our circular dependencies (es modules does not seem to have this issue). On top of this the mocha on-the-fly ts compilation is broken even after the refactor, so the tests are run against the bundle now until we can fix it or switch to a more modern test framework.

We still have a circular dependency between the chain types and the serializer component but resolving that would require a much more  extensive refactor and would probably result in a less convenient external API.